+++ my-app/src/app/layout.tsx
@@ -1,0 +1,29 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import "./globals.css";
+
+const inter = Inter({
+  variable: "--font-inter",
+  subsets: ['latin']
+})
+
+export const metadata: Metadata = {
+  title: "Receitas",
+  description: "Site de receitas",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="pt-BR">
+      <body
+        className={`${inter.variable}antialiased`}
+      >
+        {children}
+      </body>
+    </html>
+  );
+}
