+++ my-app/src/app/page.tsx
@@ -1,0 +1,8 @@
+
+export default function Home() {
+  return (
+    <div>
+      <h1>Tela inicial</h1>
+    </div>
+  );
+}
