--- receitas/src/app/globals.css
@@ -1,2 +1,0 @@
-@import "tailwindcss";
-
