--- receitas/src/components/Footer/index.tsx
@@ -1,9 +1,0 @@
-export default function Footer(){
-    return(
-    <footer className="bg-white py-4 text-black">
-        <div className="container mx-auto text-center">
-            <p>©2025 Receitas Deliciosas. Todos os direitos reservados</p>
-        </div>
-    </footer>
-    )
-}
